incorporate more PR feedback
