+++ docs/docs/supreme.md
@@ -688,9 +688,12 @@ The Supreme KMP crypto provider implements the following key derivation function
 Usage is the same across implementations:
 
 1. Instantiate a `KDF` implementation using algorithm-specific parameters as per the respective RFCs. These are:
-    * HKDF comes predefined for the SHA-1 and SHA-2 family of hash functions as `HKDF.SHA1`..`HKDF.SHA512`. Pass `info` bytes to obtain a fully instantiated `WithInfo` object: `HKDF.SHAXXX(info)` 
-    * PBKDF2 comes predefined for HMAC based on the SHA-1 and SHA-2 family of hash functions as `PBKDF2.HMAC_SHA1`..`PBKDF2.HMAC_SHA512`. Pass the number of `iterations` is required to obtain a `WithIterations` object: `PBKDF2.SHAXXX(iterations)`
-    * An scrypt instance can be configured as desired: `SCrypt(cost, parallelization, blockSize)`.
+    * HKDF comes predefined for the SHA-1 and SHA-2 family of hash functions as `HKDF.SHA1`..`HKDF.SHA512`. Pass `info` bytes to obtain a fully instantiated `WithInfo` object:  
+    `HKDF.SHAXXX(info = "CONSTANT".encodeToByteArray())` 
+    * PBKDF2 comes predefined for HMAC based on the SHA-1 and SHA-2 family of hash functions as `PBKDF2.HMAC_SHA1`..`PBKDF2.HMAC_SHA512`. Pass the number of `iterations` is required to obtain a `WithIterations` object:  
+    `PBKDF2.SHAXXX(iterations = aNumberInTheTenThousands)`
+    * An scrypt instance can be configured as desired:  
+    `SCrypt(cost, parallelization, blockSize)`.
 2. Invoke `deriveKey(salt, inputKeyMaterial, derivedKeyLength)` to obtain a derived key of length `derivedKeyLength` based on `inputKeyMaterial` and the provided `salt`.
 
 In line with other APIs, `deriveKey` returns a `KmmResult` indicating either success or failure.
